iklan textarea align& confirm on delete

+++ advert_edit.php
@@ -120,6 +120,14 @@ if (isset($_POST['title'])){
                 flex-direction: column;
             }
 
+            textarea{
+                text-align:left;
+            }
+
+            textarea:focus{
+                text-align:left;
+            }
+
             table, thead, tbody, th, td, tr { 
             display: flex;
             flex-direction: column;
@@ -293,7 +301,7 @@ if (isset($_POST['title'])){
                 echo ("<td>".$row['slot']."</td>");
                 echo ("<td><img class=\"logo\" src=\"iklan/".$row['image']."\" alt=\"test\"></td>");
                 echo ('<td><a href="advert_edit.php?adid='.$row['adid'].'">Edit /</a>');          
-                echo ('<a href="advert_delete.php?adid='.$row['adid'].'">Delete</a></td>');             
+                echo ('<a href="advert_delete.php?adid='.$row['adid'].'" onclick="return confirm(\'Hapus Iklan?\')">Delete</a></td>');             
                 echo ("</tr>");
             }
             ?>
